+++ framework/modules/addressbook/controllers/addressController.php
@@ -348,7 +348,7 @@ class addressController extends expController {
 
         $line_end = ini_get('auto_detect_line_endings');
         ini_set('auto_detect_line_endings',TRUE);
-        $checkhandle = fopen($file->path, "r");
+        $checkhandle = fopen($file->path, "rb");
         if ($this->params['type_of_address'][0] == 'am') {
             $checkdata = fgetcsv($checkhandle, 10000, "\t");
             $fieldCount = count($checkdata);
@@ -386,7 +386,7 @@ class addressController extends expController {
         //exit();
         $line_end = ini_get('auto_detect_line_endings');
         ini_set('auto_detect_line_endings',TRUE);
-        $handle = fopen($file->path, "r");
+        $handle = fopen($file->path, "rb");
         $data = fgetcsv($handle, 10000, ",");
         //eDebug($data);
         $dataset = array();