+++ framework/plugins/bootstrap5/function.icon.php
@@ -75,6 +75,10 @@ if (!function_exists('smarty_function_icon')) {
 
         if (!isset($params['int'])) $params['int'] = $loc->int;
 
+        if (!isset($params['action'])) {
+            $params['action'] = '';
+        }
+
         // attempt to translate the alt, text, & title
         if (!empty($params['alt'])) {
             $params['alt'] = gt($params['alt']);