+++ CHANGELOG.md
@@ -30,6 +30,8 @@ Version 2.7.1
   - initial support for PHP v8.2.x
 - add trash feature to file manager
 - add/fix tinymce5 as an editor option
+- add option to allow valid smtp user required to send mail
+- add sitemap.xml support for system user
 - fix maintenance view selected by theme framework
 - fix optional clock display
 - fixes and tweaks many issues in previous versions